+++ src/parse/util.rs
@@ -25,19 +25,17 @@ pub fn attr_eq(attr: &Attribute, name: &str) -> bool {
     }
 }
 
-/// checks that a function signature
-///
-/// - has no bounds (like where clauses)
-/// - is not `async`
-/// - is not `const`
-/// - is not `unsafe`
-/// - is not generic (has no type parametrs)
-/// - is not variadic
-/// - uses the Rust ABI (and not e.g. "C")
+// checks that a function signature
+//
+// - has no bounds (like where clauses)
+// - is not `const`
+// - is not `unsafe`
+// - is not generic (has no type parametrs)
+// - is not variadic
+// - uses the Rust ABI (and not e.g. "C")
 pub fn check_fn_signature(item: &ItemFn) -> bool {
     item.vis == Visibility::Inherited
         && item.sig.constness.is_none()
-        && item.sig.asyncness.is_none()
         && item.sig.abi.is_none()
         && item.sig.unsafety.is_none()
         && item.sig.generics.params.is_empty()
@@ -45,14 +43,13 @@ pub fn check_fn_signature(item: &ItemFn) -> bool {
         && item.sig.variadic.is_none()
 }
 
-// TODO: Fix later when dealing with external tasks
+// checks for foreign (extern) functions as above
 #[allow(dead_code)]
 pub fn check_foreign_fn_signature(item: &ForeignItemFn) -> bool {
     item.vis == Visibility::Inherited
-        // && item.constness.is_none()
-        // && item.asyncness.is_none()
-        // && item.abi.is_none()
-        // && item.unsafety.is_none()
+        && item.sig.constness.is_none()
+        // && item.abi.is_none() // TODO: check what this really means
+        && item.sig.unsafety.is_none()
         && item.sig.generics.params.is_empty()
         && item.sig.generics.where_clause.is_none()
         && item.sig.variadic.is_none()